Piece-list initialization and debuggers helper
A function is defined to update the piece-lists when a new position is
set on the board. The previous initialization function for the pieces
position in the list was removed.

A new command in the uci loop was added to print the board state on the
screen and the values of the squares in the piece-list.

Usage example:
ucinewgame
print

12x10 Chess Board:
==============================
10 │ X X X X X X X X X X │
09 │ X X X X X X X X X X │
08 │ X r n b k q b n r X │
07 │ X p p p p p p p p X │
06 │ X . . . . . . . . X │
05 │ X . . . . . . . . X │
04 │ X . . . . . . . . X │
03 │ X . . . . . . . . X │
02 │ X P P P P P P P P X │
01 │ X R N B K Q B N R X │
00 │ X X X X X X X X X X │
-1 │ X X X X X X X X X X │
   └─────────────────────
     z a b c d e f g h i

Piece List Contents:
========================================
White Kings         : 3
White Queens        : 4
White Rooks         : 0 7
White Light Bishops : 5
White Dark Bishops  : 2
White Knights       : 1 6
White Pawns         : 8 9 10 11 12 13 14 15
Black Kings         : 59
Black Queens        : 60
Black Rooks         : 56 63
Black Light Bishops : 58
Black Dark Bishops  : 61
Black Knights       : 57 62
Black Pawns         : 48 49 50 51 52 53 54 55