Add castling move and fix en passant capture
The en passant capture was fixed with the correct board width offset
in the pawn move generation.

To make the castling work, a few functions were necessary to check if
the squares in the king trajectory were in check, and the castling
rights were moved from the game state to the board scope, with new
structure to better handle the castling rights.

A bug was found when checking if the king is in check, the error steamed
from not updating the new piece positions in the list when the move was
made, leading to an incorrect board analisys.