Add pawn test file
Part of the code that tested the pawn piece was moved from board.rs
to it's own test file (pawn_tests.rs). Beside this code refactor, a new
function was added in GameState to generate moves an return the moves in
algebraic notation. This way it was possible to write tests without
having to worry about the move structure used by the internals of the
library