Add initial benchmarks
This commit adds benchmark for different game positions and in different
depths, this give us the base metric for move generation.
Along with that, some benchmarks were added to test the basic
chess-engine operations to measure responsiveness and delay.

The benchmarks can be run with:
`cargo bench`

Or we can specify which benchmark to run with:
`cargo bench -- <crate>` eg. `cargo bench -- chess_benchmarks`