Refactor castling check responsability
Here the logic that verifies if castling is possible was moved from the
piece-list implementation and put on the ChessBoard structure since it
can be used from different moving generation implementations.

The direction vector between two squares was decoupled from the bishop
and rook logic attack so that the piece-list module won't have to deal
with how the chess board is being represented.