Add time control to stop search automatically
Previously, the search would only stop after exhausting all moves up to
the target depth or when a "stop" command was received from the UCI
interface. This change introduces a simple time controller that
interrupts the search when the allotted time expires, preventing the
engine from running out of time during a match.